requirements.txt
streamlit
PyPDF2
langchain
langchain_community
langchain_core
langchain_google_genai
python-dotenv
faiss-cpu
google-generativeai
streamlit
beautifulsoup4
